+++ src/features/crm/services/index.ts
@@ -2,3 +2,4 @@ export * from './crm.service';
 export * from './tasks.service';
 export * from './visits.service';
 export * from './deepDive.service';
+export * from './priceList.service';